+++ QuizApp/QuizApp/Repositories/RepoModels.swift
@@ -54,6 +54,19 @@ extension QuizRepoModel {
 
 }
 
+extension QuizRepoModel {
+
+    init(_ model: QuizDatabaseModel) {
+        id = model.id
+        name = model.name
+        description = model.desc
+        category = CategoryRepoModel(rawValue: model.category)!
+        difficulty = DifficultyRepoModel(rawValue: model.difficulty)!
+        imageUrl = model.imageUrl
+        numberOfQuestions = model.numberOfQuestions
+    }
+}
+
 enum CategoryRepoModel: String {
 
     case sport = "SPORT"
